refactor files to avoid circular imports
