+++ nucleus/__init__.py
@@ -97,7 +97,9 @@ from .constants import (
 from .data_transfer_object.dataset_details import DatasetDetails
 from .data_transfer_object.dataset_info import DatasetInfo
 from .dataset import Dataset
-from .dataset_item import CameraParams, DatasetItem, Quaternion
+from .dataset_item import DatasetItem
+from .quaternion import Quaternion
+from .camera_params import CameraParams
 from .deprecation_warning import deprecated
 from .errors import (
     DatasetItemRetrievalError,
